+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v11.0.0 Breaking changes
 
 - Remove `Tooltip` component, use [@goodhood/components Tooltip](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Remove `Slider` component, use [@goodhood/components Slider](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 
 ## v10.0.0 Breaking changes