+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Remove `Markdown` component, use [@goodhood/components Markdown](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 - Remove `ContentHeader` component, use [@goodhood/components ContentHeader](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 - Remove `LinkHeader` component, use [@goodhood/components LinkHeader](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
+- Remove `Dots` component, use [@goodhood/components Dots](https://github.com/goodhood-eu/goodhood/tree/master/packages/components) instead
 
 ## v10.0.0 Breaking changes